Issue #1142 - move code validation logic to DefaultTermServiceProvider
