+++ fhir-path/src/main/java/com/ibm/fhir/path/function/MemberOfFunction.java
@@ -6,7 +6,6 @@
 
 package com.ibm.fhir.path.function;
 
-import static com.ibm.fhir.core.util.LRUCache.createLRUCache;
 import static com.ibm.fhir.model.util.ModelSupport.FHIR_STRING;
 import static com.ibm.fhir.path.evaluator.FHIRPathEvaluator.SINGLETON_FALSE;
 import static com.ibm.fhir.path.evaluator.FHIRPathEvaluator.SINGLETON_TRUE;
@@ -17,22 +16,14 @@ import static com.ibm.fhir.path.util.FHIRPathUtil.isCodedElementNode;
 import static com.ibm.fhir.path.util.FHIRPathUtil.isStringElementNode;
 import static com.ibm.fhir.path.util.FHIRPathUtil.isStringValue;
 import static com.ibm.fhir.path.util.FHIRPathUtil.isUriElementNode;
-import static com.ibm.fhir.term.util.ValueSetSupport.getContains;
 import static com.ibm.fhir.term.util.ValueSetSupport.getValueSet;
+import static com.ibm.fhir.term.util.ValueSetSupport.isExpandable;
+import static com.ibm.fhir.term.util.ValueSetSupport.isExpanded;
 
 import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import com.ibm.fhir.model.resource.ValueSet;
-import com.ibm.fhir.model.resource.ValueSet.Expansion;
-import com.ibm.fhir.model.resource.ValueSet.Expansion.Contains;
 import com.ibm.fhir.model.type.Code;
 import com.ibm.fhir.model.type.CodeableConcept;
 import com.ibm.fhir.model.type.Coding;
@@ -54,11 +45,6 @@ import com.ibm.fhir.term.service.FHIRTermService;
  * is used to determine whether or not to add a warning to the evaluation context if the membership check fails.
  */
 public class MemberOfFunction extends FHIRPathAbstractFunction {
-    private static final Logger log = Logger.getLogger(MemberOfFunction.class.getName());
-
-    private static final String VERSION_UNKNOWN = "<version unknown>";
-    private static final Map<String, Map<String, Set<String>>> CODE_SET_MAP_CACHE = createLRUCache(1024);
-
     @Override
     public String getName() {
         return "memberOf";
@@ -98,30 +84,29 @@ public class MemberOfFunction extends FHIRPathAbstractFunction {
         String strength = (arguments.size() == 2) ? getString(arguments.get(1)) : null;
 
         if (FHIRRegistry.getInstance().hasResource(url, ValueSet.class)) {
-            Map<String, Set<String>> codeSetMap = getCodeSetMap(url);
-            if (!codeSetMap.isEmpty()) {
+            ValueSet valueSet = getValueSet(url);
+            if (isExpanded(valueSet) || isExpandable(valueSet)) {
+                FHIRTermService service = FHIRTermService.getInstance();
                 if (element.is(Code.class)) {
                     String system = getSystem(evaluationContext.getTree().getParent(elementNode));
                     String code = element.as(Code.class).getValue();
-                    if (contains(codeSetMap, system, null, code)) {
+                    if (service.validateCode(valueSet, system, null, code)) {
                         return SINGLETON_TRUE;
                     }
                 } else if (element.is(Coding.class)) {
                     Coding coding = element.as(Coding.class);
-                    if (contains(codeSetMap, coding)) {
+                    if (service.validateCode(coding)) {
                         return SINGLETON_TRUE;
                     }
                 } else if (element.is(CodeableConcept.class)) {
                     CodeableConcept codeableConcept = element.as(CodeableConcept.class);
-                    for (Coding coding : codeableConcept.getCoding()) {
-                        if (contains(codeSetMap, coding)) {
-                            return SINGLETON_TRUE;
-                        }
+                    if (service.validateCode(valueSet, codeableConcept)) {
+                        return SINGLETON_TRUE;
                     }
                 } else {
                     // element.is(FHIR_STRING) || element.is(Uri.class)
                     String value = element.is(FHIR_STRING) ? element.as(FHIR_STRING).getValue() : element.as(Uri.class).getValue();
-                    if (contains(codeSetMap, null, null, value)) {
+                    if (service.validateCode(valueSet, null, null, value)) {
                         return SINGLETON_TRUE;
                     }
                 }
@@ -146,63 +131,6 @@ public class MemberOfFunction extends FHIRPathAbstractFunction {
         return SINGLETON_FALSE;
     }
 
-    private boolean contains(Map<String, Set<String>> codeSetMap, Coding coding) {
-        String system = (coding.getSystem() != null) ? coding.getSystem().getValue() : null;
-        String version = (coding.getVersion() != null) ? coding.getVersion().getValue() : null;
-        String code = (coding.getCode() != null) ? coding.getCode().getValue() : null;
-        return contains(codeSetMap, system, version, code);
-    }
-
-    /**
-     * Determine whether the provided code is in the codeSet associated with the provided system and version.
-     *
-     * <p>If the system and version are non-null, then they are concatenated to form a key into the codeSetMap. If
-     * not found, then the system is concatenated with the "VERSION_UNKNOWN" constant (in cases where the expanded
-     * value set did not have a version available during the expansion). If only the system is non-null, then the
-     * codeSetMap keys are checked for startsWith(system). Finally, if both system and version are null, map keys
-     * are ignored and the values of the map are checked directly.
-     *
-     * @param codeSetMap
-     *     the code set map
-     * @param system
-     *     the system of the focal coded element (can be null)
-     * @param version
-     *     the version of the focal coded element (can be null)
-     * @param code
-     *     the code used in the membership check
-     * @return
-     *     true if a codeSet is found and the provided code is a member of that codeSet, false otherwise
-     */
-    private boolean contains(Map<String, Set<String>> codeSetMap, String system, String version, String code) {
-        if (system != null && version != null) {
-            Set<String> codeSet = codeSetMap.get(system + "|" + version);
-            if (codeSet != null) {
-                if (codeSet.contains(code)) {
-                    return true;
-                } else {
-                    codeSet = codeSetMap.get(system + "|" + VERSION_UNKNOWN);
-                    if (codeSet != null) {
-                        return codeSet.contains(code);
-                    }
-                }
-            }
-        } else if (system != null) {
-            String prefix = system + "|";
-            for (String key : codeSetMap.keySet()) {
-                if (key.startsWith(prefix)) {
-                    return codeSetMap.get(key).contains(code);
-                }
-            }
-        } else {
-            for (Set<String> codeSet : codeSetMap.values()) {
-                if (codeSet.contains(code)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
     /**
      * Get a URI-typed child node of the input parameter with name "system".
      *
@@ -221,34 +149,5 @@ public class MemberOfFunction extends FHIRPathAbstractFunction {
             }
         }
         return null;
-    }
-
-    private Map<String, Set<String>> getCodeSetMap(String url) {
-        return CODE_SET_MAP_CACHE.computeIfAbsent(url, k -> computeCodeSetMap(getValueSet(url)));
-    }
-
-    private Map<String, Set<String>> computeCodeSetMap(ValueSet valueSet) {
-        try {
-            ValueSet expanded = FHIRTermService.getInstance().expand(valueSet);
-            if (expanded == null || expanded.getExpansion() == null) {
-                return Collections.emptyMap();
-            }
-            Map<String, Set<String>> codeSetMap = new LinkedHashMap<>();
-            Expansion expansion = expanded.getExpansion();
-            for (Contains contains : getContains(expansion)) {
-                String system = (contains.getSystem() != null) ? contains.getSystem().getValue() : null;
-                String version = (contains.getVersion() != null && contains.getVersion().getValue() != null) ? contains.getVersion().getValue() : VERSION_UNKNOWN;
-                String code = (contains.getCode() != null) ? contains.getCode().getValue() : null;
-                if (system != null && code != null) {
-                    codeSetMap.computeIfAbsent(system + "|" + version, k -> new LinkedHashSet<>()).add(code);
-                }
-            }
-            return codeSetMap;
-        } catch (Exception e) {
-            String url = (valueSet.getUrl() != null) ? valueSet.getUrl().getValue() : "<no url>";
-            String version = (valueSet.getVersion() != null) ? valueSet.getVersion().getValue() : "<no version>";
-            log.log(Level.WARNING, String.format("Unable to expand value set with url: %s and version: %s", url, version), e);
-        }
-        return Collections.emptyMap();
     }
 }
