+++ fhir-term/src/main/java/com/ibm/fhir/term/spi/FHIRTermServiceProvider.java
@@ -10,6 +10,7 @@ import java.util.Set;
 
 import com.ibm.fhir.model.resource.CodeSystem.Concept;
 import com.ibm.fhir.model.resource.ValueSet;
+import com.ibm.fhir.model.type.CodeableConcept;
 import com.ibm.fhir.model.type.Coding;
 import com.ibm.fhir.model.type.code.ConceptSubsumptionOutcome;
 
@@ -68,6 +69,22 @@ public interface FHIRTermServiceProvider {
     boolean validateCode(Coding coding);
 
     /**
+     * Indicates whether the given code is a member of the specified value set
+     *
+     * @param valueSet
+     *     the value set
+     * @param system
+     *     the system
+     * @param version
+     *     the version
+     * @param code
+     *     the code
+     * @return
+     *     true if the given code is a member of the specified alue set, false otherwise
+     */
+    boolean validateCode(ValueSet valueSet, String system, String version, String code);
+
+    /**
      * Indicates whether the given coding is a member of the specified value set
      *
      * @param valueSet
@@ -78,4 +95,16 @@ public interface FHIRTermServiceProvider {
      *     true if the given coding is a member of the specified value set, false otherwise
      */
     boolean validateCode(ValueSet valueSet, Coding coding);
+
+    /**
+     * Indicates whether the given codeable concept contains a coding that is a member of the specified value set
+     *
+     * @param valueSet
+     *     the value set
+     * @param codeableConcept
+     *     the codeable concept
+     * @return
+     *     true if the given codeable concept contains a coding that is a member of the specified value set
+     */
+    boolean validateCode(ValueSet valueSet, CodeableConcept codeableConcept);
 }
